// src/app/[locale]/(main)/(profile-settings)/account/business/types.ts
export interface IProfileBusiness {
    user: string
    license_number: string | null
    issuing_authority: string | null
    document_type: string | null
    country: string | null
    business_address: string | null
    tax_id_number: string | null
    licence_file: string | null
    tax_file: string | null
    company_name: string | null
    company_image: string | null
    company_description: string | null
    purpose_of_business_activity: string | null
    business_type: string | null
}